When there is no last date, use 30.years ago:
Some jobs can never be enqueued because its scheduler is stopped,
now this is captured, when there is not last execution I
will return last date 30 years ago.

This generate other issue, if job has not been enqueued yet this
can generate a false positive, but this is a problem smaller than
never alert when scheduler is not working.